[AuthZ] Add support to determinte role is subset of another
For generating API keys/tokens with restricted role for a given
user we need to determine if the provided scope for the token is
subset of the role for the current logged in user.

Create API key request will carry role descriptors and we need to
verify if these are the subset of the current logged in user.
These role descriptors are stored along with the API key so we can
authorize the logins in future.

Usually we would build `Role` from the role descriptors and do a
`isSubsetOf` another role checks, but when DLS queries are
present this becomes problematic. If it is an exact match for the
queries and they are subset we can say the role is a subset, but
when they do not match we have taken an approach to combine the
DLS queries with the queries from the base role. This will
eliminate the documents which match base role queries and the
queries as specified in the role descriptor for the API key.

The problem has two parts one is to check whether the given
`Role` is a subset of another role and the other one to modify
the role descriptor if required. The check
would return either YES when the role is a subset, NO when it is
not a subset or MAYBE when the role is a subset in every other aspect
BUT DLS queries.
Once we determine the result for the subset check, in case of MAYBE
we need to find the role descriptors (the ones which match and need
modifications) and then combine the query. We combine the queries
with a bool query `must` (allow the documents that match the
all the clauses).