refactor cluster permission
This commit refactors cluster permission by introducing
a builder to build cluster permission using `ClusterPrivilege`
and/or `ConfigurableClusterPrivilege`.
This lays the groundwork for moving the automaton based
The checks are performed by the implementations of `PermissionCheck`.

With this change the cluster permission has an intimate knowledge
of *ClusterPrivilege since it extracts the predicates and uses
it to build the permission.
But in the next refactoring this will change such that the *ClusterPrivilege
will