Deprecate "create" privilege in favor of "create_doc"
This commit deprecates "create" index privilege in favor of
"create_doc" privilege.
IndexPrivilege now captures the information like if it is deprecated or
not with alternatives and this information is used in the response for
built-in privileges. This API can be used by Kibana or any other
automation tools to know about the deprecated index privilege names.

Whenever the role descriptor is parsed (from file, API invocations, during
authorization), when building the IndicesPrivilege this will log the
usage of deprecated privilege names.

Note: This will just log depecation warning messages but we do not have
any migration tool to assist user.